move ProcessHostnameWithInfraDNS out of sceNet
